# Java: Rewrite handling of ORTERecvInfo structure in middle-code