--- a/orte/include/jorte/jorte_typedefs_defines.h
+++ b/orte/include/jorte/jorte_typedefs_defines.h
@@ -27,7 +27,9 @@ typedef struct {
   jobject        rinfo;
   jobject        msg;
   jobject        obj_buf; // byte buffer object
-  void          *info_buf; // RecvInfo buffer object
+  void          *info_buf; // RecvInfo buffer pointer
+  jmethodID  mid_read;
+  jmethodID  mid_callback;
   CDR_Endianness cur_endian;
 } JORTECallbackContext_t;