… orte/libjorte/setRecvInfo.c