Switch from Thumb to ARM mode for better performance.