@wentasah 2015-03-26
On Wed, Mar 25 2015, Martin Vajnar wrote:
> Switch from Thumb to ARM mode for better performance.

Have you measured how this change influence the performance?