@Vajnar 2015-04-08
Rewrite part of the 'middle' subscriber callback function, that is responsible for preparing the 'Received Information' class.

Previously the underlying C structure was parsed every-time new issue arrived and transformed into corresponding Java fields. Now, upon first invocation of the callback function (that is when first issue is processed) a direct ByteBuffer is allocated by the JVM. The callback function then performs only a non-expensive memcpy("Java ByteBuffer", "C struct ORTERecvInfo" , sizeof(ORTERecvInfo)) every-time new issue is received. The information stored in the ByteBuffer is then processed only when requested.