@Vajnar 2015-04-08
Drop the switch from Thumb to ARM for now as this callback commit is more important and GitHub doesn't allow me to create a new pull request because of the pending one.