@Vajnar 2015-04-16
Speed-up the subscriber callback by caching method IDs and by moving away run-once code.