@Vajnar 2015-04-19
Finally found time to do some performance comparison. I compared the state without this whole patch-set and with it.


Reception of 10,000 issues.

PC, Intel C2Q Q9550 @ 2GHz, OpenJDK 7
Before: 2.01 s
Now:     1.68 s

SE Xperia Ray, Android 4.1
Before (Thumb):  10.95 s
Before (ARM):      10.90 s
Now (Thumb):       9.43 s
Now (ARM):           9.53 s

The performance of Thumb-2 was a little bit better than ARM before this patch-set, but now the Thumb-2 mode outperforms the ARM mode.