@wentasah 2015-04-20
OK but AFAICS the original performance patch is no longer in the pull request so I cannot easily merge it. Please create a new pull request and put your benchamrk results into the commit message.